Resolved bug in spider movement, added offset movement capability.

+++ install/gwpspider_interfaces/include/gwpspider_interfaces/gwpspider_interfaces/srv/detail/set_bus_watchdog__struct.h
@@ -20,7 +20,7 @@ extern "C"
 /// Struct defined in srv/SetBusWatchdog in the package gwpspider_interfaces.
 typedef struct gwpspider_interfaces__srv__SetBusWatchdog_Request
 {
-  float value;
+  int32_t value;
 } gwpspider_interfaces__srv__SetBusWatchdog_Request;
 
 // Struct for a sequence of gwpspider_interfaces__srv__SetBusWatchdog_Request.
